requirements.txt
flask
flask_cors
dotenv
requests
pyjwt
firebase-admin
argon2-cffi
Flask-Mail
pytest